全コンポーネントを分割した
