+++ pages/index.tsx
@@ -1,5 +1,5 @@
 import Home from 'src/pages/Home'
 
 export default function IndexPage() {
-  return <Home />
+  return <Home></Home>
 }
