+++ pages/news.tsx
@@ -1,8 +1,5 @@
-import { compareDesc, format, isDate, isValid } from 'date-fns'
-import MainLayout from 'src/components/MainLayout'
-import { Section } from 'src/pages/Home/elements'
-import { StyledLink as StyledLink_ } from 'src/utils/next/elements'
-import styled from 'styled-components'
+import { isValid } from 'date-fns'
+import News from 'src/pages/News'
 import { getPostsAll, MarkdownBaseMeta, MarkdownData } from './api/markdowns'
 
 interface Props {
@@ -31,36 +28,9 @@ export default function NewsPage({ postsStr }: Props) {
       }
     })
     .filter(notUndefined)
-  const sortedMetaList = filterdMetaList.sort((a, b) => compareDesc(a.date, b.date))
 
-  return (
-    <MainLayout>
-      <Section>
-        {sortedMetaList.map(({ date, caption, path }) => (
-          <MetaItem key={path}>
-            <Hiduke>{format(date, 'yyyy/MM/dd')}</Hiduke>
-            <StyledLink href={path}>{caption}</StyledLink>
-          </MetaItem>
-        ))}
-      </Section>
-    </MainLayout>
-  )
+  return <News metaList={filterdMetaList}></News>
 }
-
-const StyledLink = styled(StyledLink_)`
-  text-decoration: underline;
-`
-
-const Hiduke = styled.time`
-  font-weight: bold;
-  font-family: 'novecentosans', sans-serif;
-  margin-right: 8px;
-`
-
-const MetaItem = styled.div`
-  color: #333;
-  margin-bottom: 24px;
-`
 
 export async function getStaticProps() {
   const posts = getPostsAll()
